Move get_container_status to crd process proxy
